Make environment variables access platform-specific
